requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pillow==11.3.0
pillow-heif==1.1.0
python-multipart==0.0.20

